redux initial set up
