+++ package.json
@@ -16,6 +16,9 @@
     "react-scripts": "3.4.1",
     "react-spotify-login": "^0.9.1",
     "react-spotify-player": "^1.0.4",
+    "redux": "^4.0.5",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0",
     "request": "^2.88.2"
   },
   "scripts": {
