+++ src/App.js
@@ -1,8 +1,7 @@
 import React, { useState, useEffect } from "react"
-import logo from "./logo.svg"
 import "./App.css"
 import { createNewRoom, getRoom, getCurrentRoomData } from "./firebase/firebase"
-
+import { spotfityLogin, getNewToken, getMyData } from "./spotifyLogin"
 import axios from "axios"
 import queryString from "querystring"
 
@@ -22,64 +21,8 @@ function App() {
   const [mySpotifyData, setMySpotifyData] = useState()
 
   useEffect(() => {
-    const code = new URLSearchParams(window.location.search).get("code")
-    if (code !== null) {
-      const accessForm = queryString.stringify({
-        grant_type: "authorization_code",
-        code,
-        redirect_uri: redirectUri,
-      })
-      // base64 encode auth data
-      const auth = btoa(`${clientId}:${clientSecret}`)
-      axios
-        .post("https://accounts.spotify.com/api/token", accessForm, {
-          headers: {
-            "content-type": "application/x-www-form-urlencoded;charset=utf-8",
-            Authorization: `Basic ${auth}`,
-          },
-        })
-        .then((res) => {
-          console.log(res)
-          // removes 'code' query param to clean up URL
-          window.history.replaceState(null, null, window.location.pathname)
-          setToken(res.data.access_token)
-          setRefreshToken(res.data.refresh_token)
-          console.log(token)
-        })
-        .catch((err) => console.log(err))
-    }
-  })
-
-  const getNewToken = () => {
-    const accessForm = queryString.stringify({
-      grant_type: "refresh_token",
-      refresh_token: refreshToken,
-    })
-    const auth = btoa(`${clientId}:${clientSecret}`)
-    axios
-      .post("https://accounts.spotify.com/api/token", accessForm, {
-        headers: {
-          "content-type": "application/x-www-form-urlencoded;charset=utf-8",
-          Authorization: `Basic ${auth}`,
-        },
-      })
-      .then((res) => {
-        console.log(res)
-        setToken(res.data.access_token)
-      })
-  }
-  setTimeout(getNewToken, 3600 * 1000)
-
-  const getMyData = () => {
-    if (token) {
-      fetch("https://api.spotify.com/v1/me", {
-        headers: { Authorization: "Bearer " + token },
-      })
-        .then((res) => res.json())
-        .catch((err) => console.log(err))
-        .then((data) => setMySpotifyData(data))
-    }
-  }
+    spotfityLogin(setToken, setRefreshToken)
+  }, [])
 
   const getSampleData = async (token, episodeId) => {
     try {
@@ -92,26 +35,33 @@ function App() {
         }
       )
       console.log("EPISODE", episode)
-      return episode
+      setEpisodeUrl(episode)
     } catch (err) {
       console.log(err)
     }
   }
 
   const buttonClick = () => {
-    createNewRoom({
-      name: "room11",
-      password: "123",
-      users: [
-        { name: "Alona", accessToken: token, email: "some email" },
-        { name: "Justin", accessToken: "token", email: "email" },
-      ],
-      currentPodcast: { apiData: "" },
-    })
+    if (mySpotifyData && token) {
+      console.log("sending data")
+      createNewRoom({
+        name: "room11",
+        password: "123",
+        users: [
+          {
+            name: mySpotifyData.display_name,
+            accessToken: token,
+            refreshToken: refreshToken,
+          },
+          { name: "Justin", accessToken: "token" },
+        ],
+        currentPodcast: { apiData: "" },
+      })
+    }
+
     getRoom("room11", "123").then((res) => getCurrentRoomData(res))
     getSampleData(token, "1oLdBqEIgphJN3O6ULyw4T")
-    getMyData()
-    getNewToken()
+    //getMyData(token, setMySpotifyData)
   }
 
   return (
