+++ src/redux/store.js
@@ -1,0 +1,3 @@
+import { createStore, applyMiddleware } from "redux"
+import { createLogger } from "redux-logger"
+import thunkMiddleware from "redux-thunk"
